Added example executable
The program attached is an example of how this code can be used. It uses OpenGL and Pygame to render the attached height_map.png. It generates a load of frames in the same directory and then renders it into a video (Requires FFMPEG).
If you want to use your own height map, replace this one with a 100x100 or similar sized height map with the same name.